+++ R/check.r
@@ -14,7 +14,7 @@ check <- function(x, ...) UseMethod("check")
 ##' @param participant.age.column the column in the \code{participants} data frame containing participants' age
 ##' @param country.column the column in the \code{participants} data frame containing the country in which the participant was queried
 ##' @param year.column the column in the \code{participants} data frame containing the year in which the participant was queried
-##' @param contact.age.column the column in the \code{contacts} data frame containing contacts' age
+##' @param contact.age.column the column in the \code{contacts} data frame containing the age of contacts; if this does not exist, at least columns "..._exact", "..._est_min" and "..._est_max" must (see the \code{estimated.contact.age} option in \code{\link{contact_matrix}})
 ##' @param ... ignored
 ##' @return invisibly returns a character vector of the relevant columns
 ##' @examples
@@ -28,29 +28,57 @@ check.survey <- function(x, columns=FALSE, quiet=FALSE, error=FALSE, id.column="
     if (!is.data.frame(x$participants) || !is.data.frame(x$contacts))
         stop("The 'participants' and 'contacts' elements of 'x' must be data.frames")
 
+    x <- clean(x, sample.contact.age = TRUE)
+
     success <- TRUE
     if (!missing(columns)) {
         if (!(id.column %in% colnames(x$participants) &&
               id.column %in% colnames(x$contacts)))
         {
-            error_func("Error: id.columns '", id.column, "' does not exist in both the ",
+            error_func("id.columns '", id.column, "' does not exist in both the ",
                  "participants and contacts data frames")
             success <- FALSE
         }
 
         if (!(participant.age.column %in% colnames(x$participants)))
         {
-            error_func("Error: participant age column '", participant.age.column, "' does not exist ",
+            error_func("participant age column '", participant.age.column, "' does not exist ",
                  "in the participant data frame")
             success <- FALSE
         }
 
         if (!(contact.age.column %in% colnames(x$contacts)))
         {
-            error_func("Error: contact age column '", contact.age.column, "' does not exist ",
-                 "in the contact data frame")
+            exact.column <- paste(contact.age.column, "exact", sep="_")
+            min.column <- paste(contact.age.column, "est_min", sep="_")
+            max.column <- paste(contact.age.column, "est_max", sep="_")
+
+            if (!((exact.column %in% colnames(x$contacts)) ||
+                  (min.column %in% colnames(x$contacts) && max.column %in% colnames(x$contacts))))
+            {
+                error_func("contact age column '", contact.age.column,
+                           "' or columsn to estimate contact age ('", exact.column, "' or '",
+                           min.column, "' and '", max.column, "') do not exist in the contact data frame")
+            }
             success <- FALSE
         }
+
+        if (!(country.column %in% colnames(x$participants)))
+        {
+            error_func("country column '", country.column, "' does not exist ",
+                 "in the participant data frame")
+            success <- FALSE
+        } else
+        {
+            suppressWarnings(corrected_countries <-
+                               countrycode(x$participants[, get(country.column)],
+                                           "country.name", "country.name"))
+          failed_countries <- corrected_countries[is.na(corrected_countries)]
+          if (length(failed_countries) > 0)
+          {
+            stop("Mis-spelled countries: could not find ", paste(failed_countries, sep=", "), ".")
+          }
+         }
     }
 
     if (!quiet) {