+++ R/contact_matrix.r
@@ -12,12 +12,13 @@
 ##' @param counts whether to return counts (instead of means)
 ##' @param symmetric whether to make matrix symmetric
 ##' @param split whether to split the number of contacts and assortativity
-##' @param missing.participant.age if set to "remove", participants without age information are removed; if set to "keep", participants with missing age are kept and treated as a separate age group
-##' @param missing.contact.age if set to "sample", contacts without age information are sampled from all the contacts of participants of the same age group; if set to "remove", participants that that have contacts without age information are removed; if set to "keep", contacts with missing age are kept and treated as a separate age group
+##' @param estimated.contact.age if set to "mean" (default), people whose ages are given as a range (in columns named "..._est_min" and "..._est_max") but not exactly (in a column named "..._exact") will have their age set to the mid-point of the range; if set to "sample", the age will be sampled from the range; if set to "missing", age ranges will be treated as missing
+##' @param missing.participant.age if set to "remove" (default), participants without age information are removed; if set to "keep", participants with missing age are kept and treated as a separate age group
+##' @param missing.contact.age if set to "remove" (default), participants that that have contacts without age information are removed; if set to "sample", contacts without age information are sampled from all the contacts of participants of the same age group; if set to "keep", contacts with missing age are kept and treated as a separate age group
 ##' @param weights columns that contain weights
 ##' @param weigh.dayofweek whether to weigh the day of the week (weight 5 for weekdays ans 2 for weekends)
 ##' @param quiet if set to TRUE, output is reduced
-##' @param ... further arguments to pass to, \code{\link{check}} and \code{\link{pop_age}} (especially column names)
+##' @param ... further arguments to pass to \code{\link{get_survey}}, \code{\link{check}} and \code{\link{pop_age}} (especially column names)
 ##' @return a list of sampled contact matrices, and the underlying demography of the surveyed population
 ##' @importFrom stats xtabs runif median
 ##' @importFrom utils data globalVariables
@@ -30,7 +31,7 @@
 ##' data(polymod)
 ##' contact_matrix(polymod, countries = "United Kingdom", age.limits = c(0, 1, 5, 15))
 ##' @author Sebastian Funk
-contact_matrix <- function(survey, countries=c(), survey.pop, age.limits, filter, n = 1, bootstrap, counts = FALSE, symmetric = FALSE, split = FALSE, missing.participant.age = c("remove", "keep"), missing.contact.age = c("sample", "remove", "keep"), weights = c(), weigh.dayofweek = FALSE, quiet = FALSE, ...)
+contact_matrix <- function(survey, countries=c(), survey.pop, age.limits, filter, n = 1, bootstrap, counts = FALSE, symmetric = FALSE, split = FALSE, estimated.contact.age=c("mean", "sample", "missing"), missing.participant.age = c("remove", "keep"), missing.contact.age = c("remove", "sample", "keep"), weights = c(), weigh.dayofweek = FALSE, quiet = FALSE, ...)
 {
     ## circumvent R CMD CHECK errors by defining global variables
     lower.age.limit <- NULL
@@ -42,6 +43,7 @@ contact_matrix <- function(survey, countries=c(), survey.pop, age.limits, filter
     dayofweek <- NULL
     contact.age.group <- NULL
     proportion <- NULL
+    participants <- NULL
 
     dot.args <- list(...)
     unknown.args <- setdiff(names(dot.args), union(names(formals(check.survey)), names(formals(pop_age))))
@@ -55,6 +57,7 @@ contact_matrix <- function(survey, countries=c(), survey.pop, age.limits, filter
     missing.contact.age.set <- !missing(missing.contact.age)
 
     ## read arguments
+    estimated.contact.age <- match.arg(estimated.contact.age)
     missing.participant.age <- match.arg(missing.participant.age)
     missing.contact.age <- match.arg(missing.contact.age)
 
@@ -63,15 +66,11 @@ contact_matrix <- function(survey, countries=c(), survey.pop, age.limits, filter
     ## check and get columns
     columns <- check(survey, columns=TRUE, quiet=TRUE, ...)
 
-    ## copy data
-    participants <- copy(survey$participants)
-    contacts <- copy(survey$contacts)
-
     ## if bootstrap not asked for
     if (missing(bootstrap)) bootstrap <- (n > 1)
 
     ## check if specific countries are requested (if a survey contains data from multiple countries)
-    if (length(countries) > 0 && columns[["country"]] %in% colnames(participants))
+    if (length(countries) > 0 && columns[["country"]] %in% colnames(survey$participants))
     {
         if (all(nchar(countries) == 2))
         {
@@ -82,22 +81,22 @@ contact_matrix <- function(survey, countries=c(), survey.pop, age.limits, filter
             suppressWarnings(corrected_countries <-
                                  countrycode(countries, "country.name", "country.name"))
         }
-        present_countries <- unique(as.character(participants[[columns[["country"]]]]))
+        present_countries <- unique(as.character(survey$participants[[columns[["country"]]]]))
         missing_countries <- countries[which(is.na(corrected_countries))]
         if (length(missing_countries) > 0)
         {
             stop("Survey data not found for ", paste(missing_countries, sep=", "), ".")
         }
         countries <- corrected_countries
-        participants <- participants[get(columns[["country"]]) %in% countries]
-        if (nrow(participants) == 0)
+        survey$participants <- survey$participants[get(columns[["country"]]) %in% countries]
+        if (nrow(survey$participants) == 0)
         {
             stop("No participants left after selecting countries.")
         }
     }
 
-    ## check maximum age in the data
-    max.age <- max(participants[, get(columns[["participant.age"]])], na.rm = TRUE) + 1
+    ## check maximum participant age in the data
+    max.age <- max(survey$participants[, get(columns[["participant.age"]])], na.rm = TRUE) + 1
     if (missing(age.limits))
     {
         age.limits <- c(0, seq_len(max.age-1))
@@ -109,34 +108,71 @@ contact_matrix <- function(survey, countries=c(), survey.pop, age.limits, filter
     }
 
     ## check if any filters have been requested
-    if (nrow(contacts) > 0 && !missing(filter)) {
-        missing_columns <- setdiff(names(filter), colnames(contacts))
-        if (length(missing_columns) > 0) {
-            warning("filter column(s) ", paste(missing_columns), " not found")
-            filter <- filter[names(filter) %in% colnames(contacts)]
+    if (!missing(filter)) {
+        missing_columns <- list()
+        for (table in c("participants", "contacts"))
+        {
+            if (nrow(survey[[table]]) > 0)
+            {
+                missing_columns <-
+                    c(missing_columns, list(setdiff(names(filter), colnames(survey[[table]]))))
+                ## filter contact data
+                for (column in names(filter))
+                {
+                    if (column %in% colnames(survey[[table]]))
+                    {
+                        survey[[table]] <- survey[[table]][get(column) == filter[[column]]]
+                    }
+                }
+            }
         }
-        ## filter contact data
-        for (column in names(filter)) {
-            contacts <- contacts[get(column) == filter[[column]]]
+        missing_all <- do.call(intersect, missing_columns)
+        if (length(missing_all) > 0) {
+            warning("filter column(s) ", paste(missing_all), " not found")
         }
     }
 
     if (missing.participant.age == "remove" &&
-        nrow(participants[is.na(get(columns[["participant.age"]]))]) > 0)
+        nrow(survey$participants[is.na(get(columns[["participant.age"]]))]) > 0)
     {
         if (!quiet && !missing.participant.age.set)
         {
             message("Removing participants without age information. ",
                     "To change this behaviour, set the 'missing.participant.age' option")
         }
-        participants <- participants[!is.na(get(columns[["participant.age"]]))]
+        survey$participants <- survey$participants[!is.na(get(columns[["participant.age"]]))]
+    }
+
+    ## set contact age if it's not in the data
+    if (!(columns[["contact.age"]] %in% colnames(survey$contacts)))
+    {
+        survey$contacts[, paste(columns[["contact.age"]]) := NA_integer_]
+
+        exact.column <- paste(columns[["contact.age"]], "exact", sep="_")
+        min.column <- paste(columns[["contact.age"]], "est_min", sep="_")
+        max.column <- paste(columns[["contact.age"]], "est_max", sep="_")
+
+        if (exact.column %in% colnames(survey$contacts))
+        {
+            survey$contacts[, paste(columns[["contact.age"]]) := get(exact.column)]
+        }
+        if (min.column %in% colnames(survey$contacts) &&
+            max.column %in% colnames(survey$contacts) &&
+            estimated.contact.age != "missing")
+        {
+            survey$contacts[is.na(get(columns[["contact.age"]])) & !is.na(get(min.column)) &
+                            !is.na(get(max.column)),
+                            paste(columns[["contact.age"]]) := as.integer(rowMeans(.SD)),
+                            .SDcols=c(min.column, max.column)]
+        }
     }
 
     if (missing.contact.age == "remove" &&
-        nrow(contacts[is.na(get(columns[["contact.age"]]))]) > 0)
+        nrow(survey$contacts[is.na(get(columns[["contact.age"]]))]) > 0)
     {
-        missing.age.id <- contacts[is.na(get(columns[["contact.age"]])), get(columns[["id"]])]
-        participants <- participants[!(get(columns[["id"]]) %in% missing.age.id)]
+        missing.age.id <-
+            survey$contacts[is.na(get(columns[["contact.age"]])), get(columns[["id"]])]
+        survey$participants <- survey$participants[!(get(columns[["id"]]) %in% missing.age.id)]
     }
 
     ## if split or symmetric requested, get demographic data (survey population)
@@ -160,9 +196,9 @@ contact_matrix <- function(survey, countries=c(), survey.pop, age.limits, filter
             {
                 ## neither survey population nor country names given - try to
                 ## guess country or countries surveyed from participant data
-                if (columns[["country"]] %in% names(participants))
+                if (columns[["country"]] %in% colnames(survey$participants))
                 {
-                    survey.countries <- unique(participants[, get(columns[["country"]])])
+                    survey.countries <- unique(survey$participants[, get(columns[["country"]])])
                 } else
                 {
                     warning("No 'survey.pop' or 'countries' given, and no '", columns[["country"]],
@@ -179,9 +215,10 @@ contact_matrix <- function(survey, countries=c(), survey.pop, age.limits, filter
 
                 ## check if survey data are from a specific year - in that case
                 ## use demographic data from that year, otherwise latest
-                if (columns[["year"]] %in% colnames(participants))
+                if (columns[["year"]] %in% colnames(survey$participants))
                 {
-                    survey.year <- participants[, median(get(columns[["year"]]), na.rm=TRUE)]
+                    survey.year <-
+                        survey$participants[, median(get(columns[["year"]]), na.rm=TRUE)]
                 } else
                 {
                     survey.year <- country.pop[, max(year, na.rm=TRUE)]
@@ -216,13 +253,15 @@ contact_matrix <- function(survey, countries=c(), survey.pop, age.limits, filter
 
             if (survey.representative) {
                 survey.pop <-
-                    participants[, lower.age.limit :=
-                                       reduce_agegroups(get(columns[["participant.age"]]), age.limits)]
+                    survey$participants[, lower.age.limit :=
+                                              reduce_agegroups(get(columns[["participant.age"]]),
+                                                               age.limits)]
                 survey.pop <- survey.pop[, list(population=.N), by=lower.age.limit]
                 survey.pop <- survey.pop[!is.na(lower.age.limit)]
-                if (columns[["year"]] %in% colnames(participants))
+                if (columns[["year"]] %in% colnames(survey$participants))
                 {
-                    survey.year <- participants[, median(get(columns[["year"]]), na.rm=TRUE)]
+                    survey.year <-
+                        survey$participants[, median(get(columns[["year"]]), na.rm=TRUE)]
                 }
             }
         }
@@ -241,9 +280,9 @@ contact_matrix <- function(survey, countries=c(), survey.pop, age.limits, filter
         survey.pop <- survey.pop[, list(population = sum(population)), by=lower.age.limit]
         setkey(survey.pop, lower.age.limit)
         ## re-assign lower age limits in participants
-        participants[, lower.age.limit :=
-                           reduce_agegroups(get(columns[["participant.age"]]),
-                                            survey.pop$lower.age.limit)]
+        survey$participants[, lower.age.limit :=
+                                  reduce_agegroups(get(columns[["participant.age"]]),
+                                                   survey.pop$lower.age.limit)]
         present.lower.age.limits <- unique(survey.pop$lower.age.limit)
         present.lower.age.limits <-
             present.lower.age.limits[order(present.lower.age.limits)]
@@ -255,28 +294,32 @@ contact_matrix <- function(survey, countries=c(), survey.pop, age.limits, filter
             data.table(lower.age.limit = present.lower.age.limits,
                        upper.age.limit = c(present.lower.age.limits[-1], max.age))
         ## set upper age limits and construct age groups
-        participants <- merge(participants, lower.upper.age.limits, by="lower.age.limit")
+        survey$participants <-
+            merge(survey$participants, lower.upper.age.limits, by="lower.age.limit")
         if (all(is.na(survey.pop$population))) survey.pop[, population := NULL]
     }
 
-    participants[, lower.age.limit := reduce_agegroups(get(columns[["participant.age"]]),
-                                                       age.limits[age.limits < max.age])]
-    participants[, age.group := cut(participants[, get(columns[["participant.age"]])],
-                                          breaks = union(age.limits, max.age),
-                                          right = FALSE)]
-    age.groups <- participants[, levels(age.group)]
-    age.groups[length(age.groups)] <- paste0(max(participants$lower.age.limit), "+")
-    participants[, age.group := factor(age.group, levels=levels(age.group), labels=age.groups)]
+    survey$participants[, lower.age.limit := reduce_agegroups(get(columns[["participant.age"]]),
+                                                              age.limits[age.limits < max.age])]
+    survey$participants[, age.group :=
+                              cut(survey$participants[, get(columns[["participant.age"]])],
+                                  breaks = union(age.limits, max.age),
+                                  right = FALSE)]
+    age.groups <- survey$participants[, levels(age.group)]
+    age.groups[length(age.groups)] <-
+        paste0(max(survey$participants$lower.age.limit, na.rm=TRUE), "+")
+    survey$participants[, age.group :=
+                              factor(age.group, levels=levels(age.group), labels=age.groups)]
 
-    participants[, weight := 1]
+    survey$participants[, weight := 1]
     ## assign weights to participants, to account for weekend/weekday variation
     if (weigh.dayofweek) {
-        if ("dayofweek" %in% colnames(participants))
+        if ("dayofweek" %in% colnames(survey$participants))
         {
-            participants[dayofweek %in% 1:5,
-                         weight := 5 / nrow(participants[dayofweek %in% 1:5])]
-            participants[!(dayofweek %in% 1:5),
-                         weight := 2 / nrow(participants[!(dayofweek %in% 1:5)])]
+            survey$participants[dayofweek %in% 1:5,
+                         weight := 5 / nrow(survey$participants[dayofweek %in% 1:5])]
+            survey$participants[!(dayofweek %in% 1:5),
+                         weight := 2 / nrow(survey$participants[!(dayofweek %in% 1:5)])]
         } else
         {
             warning("'weigh.dayofweek' is TRUE, but no 'dayofweek' column in the data. ",
@@ -299,16 +342,28 @@ contact_matrix <- function(survey, countries=c(), survey.pop, age.limits, filter
         if (bootstrap)
         {
             ## take a bootstrap sample from the participants
-            part.sample <- participants[sample(nrow(participants), replace = T)]
+            part.sample <- survey$participants[sample(.N, replace = T)]
         } else
         {
             ## just use all participants
-            part.sample <- copy(participants)
+            part.sample <- survey$participants
+        }
+
+        ## sample estimated contact ages
+        if (estimated.contact.age == "sample")
+        {
+            survey$contacts[!is.na(get(min.column)) & !is.na(get(max.column)),
+                            paste(columns[["contact.age"]]) :=
+                                as.integer(runif(.N,
+                                                 as.integer(min(get(min.column),
+                                                                get(max.column))),
+                                                 as.integer(max(get(min.column),
+                                                                get(max.column)))))]
         }
 
         ## gather contacts for sampled participants
         contacts.sample <-
-            data.table(merge(contacts, part.sample, by = columns[["id"]], all = F,
+            data.table(merge(survey$contacts, part.sample, by = columns[["id"]], all = F,
                              allow.cartesian = T))
 
         ## sample contacts
@@ -444,11 +499,11 @@ contact_matrix <- function(survey, countries=c(), survey.pop, age.limits, filter
         {
             if (counts) {
                 warning("'split=TRUE' does not make sense with 'counts=TRUE'; ",
-                        "will not make matrix symmetric.")
+                        "will not split the contact matrix.")
             } else if (na.present)
             {
                 warning("'split=TRUE' does not work with missing data; ",
-                        "will not make matrix symmetric\n",
+                        "will not split contact.matrix.\n",
                         warning.suggestion)
             } else
             {
@@ -486,12 +541,12 @@ contact_matrix <- function(survey, countries=c(), survey.pop, age.limits, filter
     }
 
     ## get number of participants in each age group
-    if (any(is.na(levels(part.sample$age.group)))) {
+    if (any(is.na(part.sample$age.group))) {
         useNA <- "always"
     } else {
         useNA <- "no"
     }
-    part.pop <- data.table(table(participants[, age.group], useNA = useNA))
+    part.pop <- data.table(table(survey$participants[, age.group], useNA = useNA))
     setnames(part.pop, c("lower.age.limit", "participants"))
     part.pop[, proportion := participants / sum(participants)]
 
@@ -508,4 +563,3 @@ contact_matrix <- function(survey, countries=c(), survey.pop, age.limits, filter
 
     return(return_value)
 }
-