+++ NEWS.md
@@ -1,3 +1,11 @@
+# socialmixr 0.1.2
+
+* improved downloading form Zenodo; only a single download is used now
+* improved handling of country names
+* fixed bugs in the "filter" option
+* POLYMOD data set updated to latest v1.1
+* allow user to choose how ages are estimated from age ranges
+
 # socialmixr 0.1.1
 
 * fix: dependency on R >= 3.4.0